backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
python-multipart
openai
pydantic-settings
python-dotenv
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
aiosqlite
